Fix Jest tests for ESM
